+++ projects/plugins/wpcomsh/wpcom-features/class-wpcom-features.php
@@ -420,6 +420,7 @@ class WPCOM_Features {
 	public const SOCIAL_INSTAGRAM_CONNECTION       = 'social-instagram-connection';
 	public const SOCIAL_CONNECTIONS_MANAGEMENT     = 'social-connections-management';
 	public const SOCIAL_IMAGE_AUTO_CONVERT         = 'social-image-auto-convert';
+	public const SOCIAL_EDITOR_PREVIEW             = 'social-editor-preview';
 	public const SOCIAL_MULTI_CONNECTIONS          = 'social-multi-connections';
 	public const SPACE                             = 'space';
 	public const SPACE_UPGRADED_STORAGE            = 'space-upgraded-storage';
@@ -1077,6 +1078,14 @@ class WPCOM_Features {
 			),
 		),
 		self::SOCIAL_CONNECTIONS_MANAGEMENT     => array(
+			array(
+				// This feature isn't launched yet, so we're ensuring that it's not available on any plans.
+				'before' => '1900-01-01',
+				self::WPCOM_ALL_SITES,
+				self::JETPACK_ALL_SITES,
+			),
+		),
+		self::SOCIAL_EDITOR_PREVIEW             => array(
 			array(
 				// This feature isn't launched yet, so we're ensuring that it's not available on any plans.
 				'before' => '1900-01-01',